feat: use posneg by default for shopify search inputs
