+++ .changeset/neat-cameras-tan.md
@@ -1,0 +1,5 @@
+---
+'@sajari/search-widgets': minor
+---
+
+feat: use posneg by default for shopify search inputs
