+++ src/index.ts
@@ -6,7 +6,6 @@ import SearchInput from './search-input';
 import SearchInputBinding from './search-input-binding';
 import SearchResults from './search-results';
 import withShadowRoot from './shadow-dom-wrapper';
-import TokenCheck from './token-check';
 import { WidgetType } from './types';
 
 if (!process.env.DEPLOY_SCRIPT) {
@@ -22,7 +21,6 @@ const components: Record<WidgetType, ComponentType> = {
   overlay: SearchResults as ComponentType,
   'search-input-binding': SearchInputBinding as ComponentType,
   'search-input': SearchInput as ComponentType,
-  'token-check': TokenCheck as ComponentType,
 };
 
 const attribute = 'data-widget';
